requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dataclasses-json==0.5.14
defusedxml==0.7.1
dj-rest-auth==5.0.0
Django==4.2.5
django-allauth==0.56.1
django-cors-headers==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
emoji==2.8.0
exceptiongroup==1.1.3
faiss-gpu==1.7.2
fastapi==0.99.1
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.1
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-resources==6.0.1
incremental==22.10.0
joblib==1.3.2
kombu==5.3.2
langchain==0.0.292
langsmith==0.0.37
lxml==4.9.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.28.0
overrides==7.4.0
packaging==23.1
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.3
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
PyJWT==2.8.0
pyOpenSSL==23.2.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.0
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.15
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
youtube-transcript-api==0.6.1
zope.interface==6.0
